move subset functionality into a separate module
