+++ doc/source/pyshepseg_subset.rst
@@ -1,0 +1,11 @@
+:tocdepth: 2
+
+subset
+======
+.. automodule:: pyshepseg.subset
+   :members:
+   :undoc-members:
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
